Organize `postgresql` package into several files
